js/xpconnect/tests/mochitest/test_finalizationRegistry.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test FinalizationRegistry works in the browser</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript">
      let object4 = {};

      function go() {
        SimpleTest.waitForExplicitFinish();

        // Registry with no registered objects.
        let holdings1 = [];
        let registry1 = new FinalizationRegistry(v => { holdings1.push(v); });

        // Registry with three registered objects.
        let holdings2 = [];
        let registry2 = new FinalizationRegistry(v => { holdings2.push(v); });
        registry2.register({}, 1);
        registry2.register({}, 2);
        registry2.register({}, 3);

        // Registry with registered object that is then unregistered.
        let holdings3 = [];
        let registry3 = new FinalizationRegistry(v => { holdings3.push(v); });
        let token3 = {}
        registry3.register({}, 1, token3);
        registry3.unregister(token3);

        // Registry with registered object that doesn't die.
        let holdings4 = [];
        let registry4 = new FinalizationRegistry(v => { holdings4.push(v); });
        registry4.register(object4, 1);

        // Registry observing cyclic JS data structure.
        let holdings5 = [];
        let registry5 = new FinalizationRegistry(v => { holdings5.push(v); });
        registry5.register(makeJSCycle(4), 5);

        // Registry observing cyclic DOM/JS data structure.
        let holdings6 = [];
        let registry6 = new FinalizationRegistry(v => { holdings6.push(v); });
        registry6.register(makeDOMCycle(4), 6);

        // Need to run full GC/CC/GC cycle to collect cyclic garbage through DOM
        // and JS heaps.
        SpecialPowers.DOMWindowUtils.garbageCollect();
        SpecialPowers.DOMWindowUtils.cycleCollect();
        SpecialPowers.DOMWindowUtils.garbageCollect();

        // Microtasks are run before cleanup callbacks.
        Promise.resolve().then(() => {
          is(holdings1.length, 0);
          is(holdings2.length, 0);
          is(holdings3.length, 0);
          is(holdings4.length, 0);
          is(holdings5.length, 0);
          is(holdings6.length, 0);
        });

        // setTimeout queues a task which will run after cleanup callbacks.
        setTimeout(() => {
          is(holdings1.length, 0);

          let result = holdings2.sort((a, b) => a - b);
          is(result.length, 3);
          is(result[0], 1);
          is(result[1], 2);
          is(result[2], 3);

          is(holdings3.length, 0);
          is(holdings4.length, 0);

          is(holdings5.length, 1);
          is(holdings5[0], 5);

          is(holdings6.length, 1);
          is(holdings6[0], 6);

          SimpleTest.finish();
        }, 0);
      }

      function makeJSCycle(size) {
        let first = {};
        let current = first;
        for (let i = 0; i < size; i++) {
          current.next = {};
          current = current.next;
        }
        current.next = first;
        return first;
      }

      function makeDOMCycle(size) {
        let first = {};
        let current = first;
        for (let i = 0; i < size; i++) {
          if (i % 2 === 0) {
            current.next = document.createElement("div");
          } else {
            current.next = {};
          }
          current = current.next;
        }
        current.next = first;
        return first;
      }
    </script>
  </head>
  <body onload="go()"></body>
</html>
